+++ kong/license.go
@@ -6,7 +6,7 @@ package kong
 type License struct {
 	ID        *string `json:"id,omitempty" yaml:"id,omitempty"`
 	Payload   *string `json:"payload,omitempty" yaml:"payload,omitempty"`
-	CreatedAt *int64  `json:"created_at,omitempty" yaml:"created_at,omitempty"`
+	CreatedAt *float64  `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	UpdatedAt *int64  `json:"updated_at,omitempty" yaml:"updated_at,omitempty"`
 }
